+++ src/main/java/vn/edu/fpt/hotel_management/repository/RoomRepository.java
@@ -19,34 +19,36 @@ public interface RoomRepository extends JpaRepository<Room, Integer> {
        /** Lấy tất cả phòng của một khách sạn theo hotel_id. */
        List<Room> findByHotelId(int hotelId);
 
-       /**
-        * Lọc phòng theo hotel_id và khoảng giá.
-        */
-       @Query("SELECT r FROM Room r " +
-                     "WHERE r.hotelId = :hotelId " +
-                     "AND r.price >= :minPrice " +
-                     "AND r.price <= :maxPrice " +
-                     "ORDER BY r.price ASC")
-       List<Room> findByHotelIdAndPriceRange(
-                     @Param("hotelId") int hotelId,
-                     @Param("minPrice") double minPrice,
-                     @Param("maxPrice") double maxPrice);
+    /**
+     * Lọc phòng theo hotel_id và khoảng giá.
+     */
+    @Query("SELECT r FROM Room r " +
+           "WHERE r.hotelId = :hotelId " +
+           "AND r.price >= :minPrice " +
+           "AND r.price <= :maxPrice " +
+           "ORDER BY r.price ASC")
+    List<Room> findByHotelIdAndPriceRange(
+            @Param("hotelId")   int    hotelId,
+            @Param("minPrice")  long   minPrice,
+            @Param("maxPrice")  long   maxPrice
+    );
 
-       /**
-        * Lọc phòng theo hotel_id, loại phòng và khoảng giá.
-        * Dùng r.roomType thay vì r.type để tránh JPQL reserved word.
-        */
-       @Query("SELECT r FROM Room r " +
-                     "WHERE r.hotelId = :hotelId " +
-                     "AND r.roomType IN :types " +
-                     "AND r.price >= :minPrice " +
-                     "AND r.price <= :maxPrice " +
-                     "ORDER BY r.price ASC")
-       List<Room> filterByHotelAndTypesAndPrice(
-                     @Param("hotelId") int hotelId,
-                     @Param("types") List<String> types,
-                     @Param("minPrice") double minPrice,
-                     @Param("maxPrice") double maxPrice);
+    /**
+     * Lọc phòng theo hotel_id, loại phòng và khoảng giá.
+     * Dùng r.roomType thay vì r.type để tránh JPQL reserved word.
+     */
+    @Query("SELECT r FROM Room r " +
+           "WHERE r.hotelId = :hotelId " +
+           "AND r.roomType IN :types " +
+           "AND r.price >= :minPrice " +
+           "AND r.price <= :maxPrice " +
+           "ORDER BY r.price ASC")
+    List<Room> filterByHotelAndTypesAndPrice(
+            @Param("hotelId")   int          hotelId,
+            @Param("types")     List<String> types,
+            @Param("minPrice")  long         minPrice,
+            @Param("maxPrice")  long         maxPrice
+    );
 
        /**
         * Lấy danh sách loại phòng duy nhất của một khách sạn.
